Very handy grammar debugging line.
The Pegex grammar syntax is just sugar. What matters is the compiler
grammar tree. Uncomment this line to see it.

You can see if/how your grammar changes have an impact.

A Pegex grammar tree is just a tree of regexes combined with metadata on
how they need to be applied. The Pegex::Parser simply applies this tree
to input using a fairly simple algorithm. Matches trigger actions in a
receiver class. That's basically the whole thing,